tidied server code
resolved issue #22
