+++ public/index.js
@@ -4,25 +4,6 @@ const dashboard = document.getElementById('dashboard');
 const formSubmit = document.getElementById('form-submit');
 const formContainer = document.getElementById('form-container');
 
-function postTweet() {
-  const xhr = new XMLHttpRequest();
-  const cookieMonster = document.cookie.split(';');
-  const activeCookie = cookieMonster[cookieMonster.length - 1];
-  const tweetObj = {
-   cookie: activeCookie,
-   text: tweetText.value
-  };
-  xhr.onreadystatechange = function() {
-    if (xhr.readyState == 4 && xhr.status == 200) {
-      getTweet();
-      tweetText.value = '';
-    }
-  }
-  if (tweetText.value === '') return;
-  xhr.open("POST", "/postTweet", true);
-  xhr.send(JSON.stringify(tweetObj));
-}
-
 function clearDashboard() {
   while(dashboard.firstChild){
     dashboard.removeChild(dashboard.firstChild);
@@ -43,19 +24,6 @@ function hideLoginForm () {
   dashboard.className = '';
 }
 
-function getTweet() {
-  const xhr = new XMLHttpRequest();
-  xhr.onreadystatechange = function() {
-    if (xhr.readyState == 4 && xhr.status == 200) {
-      const tweets = JSON.parse(xhr.response).reverse();
-      clearDashboard();
-      displayTweets(tweets);
-    }
-  }
-  xhr.open("GET", "/getTweet", true);
-  xhr.send();
-}
-
 function login() {
   const xhr = new XMLHttpRequest();
   const username = document.getElementById('username').value;
@@ -71,6 +39,38 @@ function login() {
   xhr.open("POST", qString, true);
   xhr.send();
   hideLoginForm();
+}
+
+function postTweet() {
+  const xhr = new XMLHttpRequest();
+  const cookieMonster = document.cookie.split(';');
+  const activeCookie = cookieMonster[cookieMonster.length - 1];
+  const tweetObj = {
+   cookie: activeCookie,
+   text: tweetText.value
+  };
+  xhr.onreadystatechange = function() {
+    if (xhr.readyState == 4 && xhr.status == 200) {
+      getTweet();
+      tweetText.value = '';
+    }
+  }
+  if (tweetText.value === '') return;
+  xhr.open("POST", "/postTweet", true);
+  xhr.send(JSON.stringify(tweetObj));
+}
+
+function getTweet() {
+  const xhr = new XMLHttpRequest();
+  xhr.onreadystatechange = function() {
+    if (xhr.readyState == 4 && xhr.status == 200) {
+      const tweets = JSON.parse(xhr.response).reverse();
+      clearDashboard();
+      displayTweets(tweets);
+    }
+  }
+  xhr.open("GET", "/getTweet", true);
+  xhr.send();
 }
 
 window.addEventListener('load', getTweet);
